Commit-2

+++ dependencyresolver/src/main/java/com/java/feed/repo/FeedGroupRepo.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FeedGroupRepo  extends JpaRepository<FeedGroup,Integer> {
+public interface FeedGroupRepo  extends JpaRepository<FeedGroup,Long> {
+
 }
